+++ offchain-modules/packages/app-monitor/src/monitor.ts
@@ -378,6 +378,7 @@ export class Monitor {
   }
 
   async observeEthEvent(): Promise<void> {
+    let continuousErrorCount = 0;
     foreverPromise(
       async () => {
         const fromBlock = this.durationConfig.eth.lastHandledBlock + 1;
@@ -403,12 +404,22 @@ export class Monitor {
           .subscribe((record) => this.onEthUnlockRecord(record));
 
         this.durationConfig.eth.lastHandledBlock = toBlock;
+        continuousErrorCount = 0;
       },
       {
         onRejectedInterval: 15000,
-        onResolvedInterval: 0,
+        onResolvedInterval: 1000,
         onRejected: (e: Error) => {
-          logger.error(`Monitor observeEthLock error:${e.stack}`);
+          continuousErrorCount++;
+          if (continuousErrorCount > 10) {
+            logger.error(
+              `Monitor observeEthLock error, continuousErrorCount: ${continuousErrorCount}, stack: ${e.stack}`,
+            );
+          } else {
+            logger.warn(
+              `Monitor observeEthLock error, continuousErrorCount: ${continuousErrorCount}, stack: ${e.stack}`,
+            );
+          }
         },
       },
     );